Created inheritence
